refactor: move to useFlightLoader
